Add post request specs and is_live to posts
This update was long overdue. The bulk of this was 20 specs for the post
resource. This was realized when I didn't have a proper test suite when
I wanted to add the is_live functionality.

The is_live functionality will allow me to prevert an anonymous user
from seeing posts that are in the works.